blank-box: Support Block API Version 2
