--- blocks/src/block/blank-box/block.js
@@ -1,220 +1,0 @@
-/**
- * Cocoon Blocks
- * @author: yhira
- * @link: https://wp-cocoon.com/
- * @license: http://www.gnu.org/licenses/gpl-2.0.html GPL v2 or later
- */
-
-import { THEME_NAME, CLICK_POINT_MSG, fullFallbackStyles } from '../../helpers';
-import { deprecated } from './deprecated';
-import { transforms } from './transforms';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import classnames from 'classnames';
-
-
-const { __ } = wp.i18n;
-const {
-  registerBlockType,
-} = wp.blocks;
-const {
-  InspectorControls,
-  InnerBlocks,
-  RichText,
-  withColors,
-  getColorClassName,
-  PanelColorSettings,
-  getFontSizeClass,
-  withFontSizes,
-  FontSizePicker,
-  ContrastChecker,
-} = wp.editor;
-const {
-  PanelBody,
-  PanelColor,
-  ColorPalette,
-  SelectControl,
-  TextareaControl,
-  ToggleControl
-} = wp.components;
-
-const {
-  Component,
-  Fragment,
-} = wp.element;
-
-const {
-  compose
-} = wp.compose;
-
-
-class CocoonBlankBoxBlock extends Component {
-  constructor() {
-    super(...arguments);
-  }
-
-  render() {
-    const {
-      attributes,
-      setAttributes,
-      mergeBlocks,
-      onReplace,
-      className,
-      backgroundColor,
-      setBackgroundColor,
-      textColor,
-      setTextColor,
-      borderColor,
-      setBorderColor,
-      fallbackBackgroundColor,
-      fallbackTextColor,
-      fallbackBorderColor,
-      fallbackFontSize,
-      fontSize,
-      setFontSize,
-    } = this.props;
-
-    const {
-      content,
-    } = attributes;
-
-    return (
-      <Fragment>
-        <InspectorControls>
-
-          <PanelColorSettings
-            title={ __( '色設定', THEME_NAME ) }
-            colorSettings={[
-              {
-                label: __( 'ボーダー色', THEME_NAME ),
-                onChange: setBorderColor,
-                value: borderColor.color,
-              },
-              {
-                label: __( '背景色', THEME_NAME ),
-                onChange: setBackgroundColor,
-                value: backgroundColor.color,
-              },
-              {
-                label: __( '文字色', THEME_NAME ),
-                onChange: setTextColor,
-                value: textColor.color,
-              },
-            ]}
-          />
-          {/*
-          <PanelBody title={ __( '文字サイズ', THEME_NAME ) } className="blocks-font-size">
-            <FontSizePicker
-              fallbackFontSize={ fallbackFontSize }
-              value={ fontSize.size }
-              onChange={ setFontSize }
-            />
-          </PanelBody>
-          */}
-        </InspectorControls>
-
-        <div
-          className={ classnames(className, {
-              'blank-box': true,
-              'block-box': true,
-              'has-text-color': textColor.color,
-              'has-background': backgroundColor.color,
-              'has-border-color': borderColor.color,
-              [backgroundColor.class]: backgroundColor.class,
-              [textColor.class]: textColor.class,
-              [borderColor.class]: borderColor.class,
-              [fontSize.class]: fontSize.class,
-          }) }
-        >
-          <InnerBlocks />
-        </div>
-
-      </Fragment>
-    );
-  }
-}
-
-registerBlockType( 'cocoon-blocks/blank-box-1', {
-
-  title: __( '白抜きボックス', THEME_NAME ),
-  icon: <FontAwesomeIcon icon={['far', 'square']} />,
-  category: THEME_NAME + '-block',
-  description: __( 'コンテンツを囲むだけのブランクボックスを表示します。', THEME_NAME ),
-  keywords: [ 'blank', 'box' ],
-
-  attributes: {
-    content: {
-      type: 'string',
-      default: CLICK_POINT_MSG,
-    },
-    backgroundColor: {
-      type: 'string',
-    },
-    customBackgroundColor: {
-      type: 'string',
-    },
-    textColor: {
-      type: 'string',
-    },
-    customTextColor: {
-      type: 'string',
-    },
-    borderColor: {
-      type: 'string',
-    },
-    customBorderColor: {
-      type: 'string',
-    },
-    fontSize: {
-      type: 'string',
-    },
-    customFontSize: {
-      type: 'string',
-    },
-  },
-
-  edit: compose([
-    withColors('backgroundColor', {textColor: 'color', borderColor: 'border-color'}),
-    withFontSizes('fontSize'),
-    fullFallbackStyles,
-  ])(CocoonBlankBoxBlock),
-  save: props => {
-    const {
-      backgroundColor,
-      customBackgroundColor,
-      textColor,
-      customTextColor,
-      borderColor,
-      customBorderColor,
-      fontSize,
-      customFontSize,
-    } = props.attributes;
-
-    const backgroundClass = getColorClassName( 'background-color', backgroundColor );
-    const textClass = getColorClassName( 'color', textColor );
-    const borderClass = getColorClassName( 'border-color', borderColor );
-    const fontSizeClass = getFontSizeClass( fontSize );
-
-
-    const className = classnames( {
-      'blank-box': true,
-      'block-box': true,
-      'has-text-color': textColor || customTextColor,
-      'has-background': backgroundColor || customBackgroundColor,
-      'has-border-color': borderColor || customBorderColor,
-      [ textClass ]: textClass,
-      [ backgroundClass ]: backgroundClass,
-      [ borderClass ]: borderClass,
-      [ fontSizeClass ]: fontSizeClass,
-    } );
-
-    return (
-      <div className={ className }>
-        <InnerBlocks.Content />
-      </div>
-    );
-  },
-
-  deprecated: deprecated,
-
-  transforms: transforms,
-});
