+++ blocks/src/block/blank-box/edit.js
@@ -1,0 +1,81 @@
+
+import { THEME_NAME } from '../../helpers';
+import { __ } from '@wordpress/i18n';
+import {
+  InspectorControls,
+  InnerBlocks,
+  withColors,
+  PanelColorSettings,
+  withFontSizes,
+  useBlockProps,
+} from '@wordpress/block-editor';
+import { Fragment } from '@wordpress/element';
+import { compose } from '@wordpress/compose';
+import classnames from 'classnames';
+
+export function BlankBoxEdit( props ) {
+  const {
+    className,
+    backgroundColor,
+    setBackgroundColor,
+    textColor,
+    setTextColor,
+    borderColor,
+    setBorderColor,
+    fontSize,
+  } = props;
+
+  const classes = classnames(className, {
+    'blank-box': true,
+    'block-box': true,
+    'has-text-color': textColor.color,
+    'has-background': backgroundColor.color,
+    'has-border-color': borderColor.color,
+    [backgroundColor.class]: backgroundColor.class,
+    [textColor.class]: textColor.class,
+    [borderColor.class]: borderColor.class,
+    [fontSize.class]: fontSize.class,
+  });
+
+  const blockProps = useBlockProps({
+    className: classes,
+  });
+
+  return (
+    <Fragment>
+      <InspectorControls>
+
+        <PanelColorSettings
+          title={ __( '色設定', THEME_NAME ) }
+          colorSettings={[
+            {
+              label: __( 'ボーダー色', THEME_NAME ),
+              onChange: setBorderColor,
+              value: borderColor.color,
+            },
+            {
+              label: __( '背景色', THEME_NAME ),
+              onChange: setBackgroundColor,
+              value: backgroundColor.color,
+            },
+            {
+              label: __( '文字色', THEME_NAME ),
+              onChange: setTextColor,
+              value: textColor.color,
+            },
+          ]}
+        />
+      </InspectorControls>
+
+      <div { ...blockProps }>
+        <InnerBlocks />
+      </div>
+
+    </Fragment>
+  );
+}
+
+export default compose([
+  withColors('backgroundColor', {textColor: 'color', borderColor: 'border-color', iconColor: 'icon-color'}),
+  withFontSizes('fontSize'),
+])(BlankBoxEdit);
