+++ blocks/src/block/blank-box/save.js
@@ -1,0 +1,42 @@
+import {
+  InnerBlocks,
+  getColorClassName,
+  getFontSizeClass,
+  useBlockProps,
+} from '@wordpress/block-editor';
+import classnames from 'classnames';
+
+export default function save({ attributes }) {
+  const {
+    backgroundColor,
+    textColor,
+    borderColor,
+    customBorderColor,
+    fontSize,
+  } = attributes;
+
+  const backgroundClass = getColorClassName( 'background-color', backgroundColor );
+  const textClass = getColorClassName( 'color', textColor );
+  const borderClass = getColorClassName( 'border-color', borderColor );
+  const fontSizeClass = getFontSizeClass( fontSize );
+
+  const className = classnames( {
+    'blank-box': true,
+    'block-box': true,
+    'has-text-color': textColor,
+    'has-background': backgroundColor,
+    'has-border-color': borderColor || customBorderColor,
+    [ textClass ]: textClass,
+    [ backgroundClass ]: backgroundClass,
+    [ borderClass ]: borderClass,
+    [ fontSizeClass ]: fontSizeClass,
+  } );
+  const blockProps = useBlockProps.save({
+    className: className,
+  });
+  return (
+    <div { ...blockProps }>
+      <InnerBlocks.Content />
+    </div>
+  );
+}
