+++ blocks/src/blocks.js
@@ -12,11 +12,13 @@ import compareVersions from 'compare-versions';
 const cocoonBlocksPro = [];
 
 //構造化したブロック
+import * as blankBox from './block/blank-box';
 import * as iconBox from './block/icon-box';
 import * as iconList from './block/icon-list';
 import * as infoBox from './block/info-box';
 
 const cocoonBlocks = [
+  blankBox,
   iconBox,
   iconList,
   infoBox,
@@ -64,7 +66,7 @@ registerCocoonBlocks();
 //ブロック
 // import './block/icon-box';
 //import './block/info-box/block.js';
-import './block/blank-box/block.js';
+//import './block/blank-box/block.js';
 import './block/sticky-box/block.js';
 import './block/tab-box/block.js';
 import './block/balloon/block.js';
